requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-extra-fields==3.7.0
sqlparse==0.4.4
typing_extensions==4.8.0
